+++ java-client/src/test/java/org/opensearch/client/opensearch/_types/query_dsl/GeoShapeQueryFieldTest.java
@@ -19,4 +19,15 @@ public class GeoShapeQueryFieldTest extends ModelTestCase {
 
         assertEquals(toJson(copied), toJson(origin));
     }
+
+    @Test
+    public void indexedShapeOnly() {
+        GeoShapeQueryField field = new GeoShapeQueryField.Builder().indexedShape(
+            i -> i.id("id").index("shapes").path("location")
+        ).build();
+
+        assertNull(field.shape());
+        assertEquals("id", field.indexedShape().id());
+        assertEquals("{\"indexed_shape\":{\"id\":\"id\",\"index\":\"shapes\",\"path\":\"location\"}}", toJson(field));
+    }
 }